requirements.txt
flask
flask-socketio
eventlet
paho-mqtt
flask